working bullet & d4rl; pre cleanup
